Сделан рефакторинг кода
